Extracted the platform specific utils to the pure platform folder
