+++ src/constraint.cpp
@@ -5,6 +5,7 @@
 // Copyright 2008-2013 Jonathan Westhues.
 //-----------------------------------------------------------------------------
 #include "solvespace.h"
+#include "ss_util.h"
 
 std::string Constraint::DescriptionString() const {
     std::string s;
